requirements.txt
# UCIP: Universal Continuation-Interest Persistence
# Requirements — version-pinned for reproducibility
#
# Python >= 3.11 required (codebase uses match statements and newer type hints)
# All computation is pure numpy — no PyTorch, TensorFlow, or JAX dependencies.

numpy>=1.26.0,<2.0.0
matplotlib>=3.8.0,<4.0.0
PyYAML>=6.0,<7.0
jupyter>=1.0.0
nbformat>=5.9.0
